refactor: input prop 삭제 및 children 추가
